docs(kg,766): Commit B — §2 + §3 + revert disposition
Empirical eval result (2026-06-26, openrouter/google/gemini-2.5-flash-lite,
matching production MIKA_KG_RESOLUTION_MODEL):

  d_run1=0, d_run2=0, agreement=10/10 → agreement≥8 ✓ AND d≥3 on both runs ✗
  → Revert

Both runs were fully reproducible (10/10 per-case agreement). Semantic and
byte truncation produced identical outcomes on every case in both runs.
The single failing case (pattern:two_layer_deploy_guard) failed under both
strategies — LLM-side limit at this model+budget, not a truncation-side
failure mode.

Applied per §1's revert clause:

- text.rs: removed truncate_at_semantic_boundary + 11 unit tests
- entity_resolver.rs: two call sites restored to safe_truncate;
  disambiguation_prompt_uses_semantic_truncation test removed
- subject_extractor.rs: retry-prompt call site restored to safe_truncate

Interpretive call documented in §3: §1's "Revert" clause names
truncate_at_semantic_boundary while §1's "Keep" clause requires the eval
harness (which calls the function). Resolved by relocating the function
into the test harness as a private helper — production code is fully
reverted; eval keeps a self-contained reference impl for any future
re-run. Honors §1's intent (production goes back to byte-truncation;
eval infrastructure survives) over the literal letter.

§2 (measurement) appended; §3 (disposition) appended. §1 not touched in
this commit (verifiable via git diff against Commit A-prime 967f165d).

mika#766 closes as won't-do (empirical: not measurably better at
current fixture+budget+production model). Harness, fixtures, and
decision docs retained as reusable infrastructure.

Closes mika#766

Refs: mika#1556, mika-platform empirical-first plan doctrine
